# web/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /records/{id}/report_assembly_result/:
    post:
      operationId: reportAssemblyResultRecord
      description: 'Update metadata for an ENA record with a screening result.


        **pk**: Record identifier'
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Record'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Record'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Record'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Record'
          description: ''
      tags:
      - records
...
components:
  schemas:
    Record:
      type: object
      properties:
        id:
          type: string
          maxLength: 1024
        accession:
          type: string
          nullable: true
          maxLength: 40
        experiment_accession:
          type: string
          nullable: true
          maxLength: 40
        run_accession:
          type: string
          nullable: true
          maxLength: 40
        sample_accession:
          type: string
          nullable: true
          maxLength: 40
        secondary_sample_accession:
          type: string
          nullable: true
          maxLength: 40
        fastq_ftp:
          type: string
          nullable: true
          maxLength: 4096
        passed_filter:
          type: boolean
          nullable: true
        filter_failed:
          type: string
          nullable: true
          maxLength: 1024
        time_fetched:
          type: string
          format: date-time
          readOnly: true
        waiting_since:
          type: string
          format: date-time
          nullable: true
        assembly_result:
          enum:
          - skipped
          - waiting
          - under consideration
          - in progress
          - fail
          - success
          type: string
        passed_screening:
          type: boolean
          nullable: true
        screening_message:
          type: string
          nullable: true
          maxLength: 4096
        assembled_genome_url:
          type: string
          nullable: true
          maxLength: 1024
        assembly_error_report_url:
          type: string
          nullable: true
          maxLength: 1024
        assembly_error_process:
          type: string
          nullable: true
          maxLength: 1024
        assembly_error_exit_code:
          type: string
          nullable: true
          maxLength: 1024
        assembly_error_stdout:
          type: string
          nullable: true
          maxLength: 1024
        assembly_error_stderr:
          type: string
          nullable: true
          maxLength: 1024
        taxon:
          type: string
      required:
      - id
      - taxon
...